cmake: rename options adding a SUSHI_ prefix
The old options names are still kept to avoid breaking existing builds.

Unfortunately, it is not trivial at all to detect if an option is
explicitly set by the user. The current behavior is to just detect if an
old option is set to a value different than the default one, and to give
precedence over the new option if the old one was defined.

Also prepared an architecture-dependent list of default options to make
macOS builds easier.